Multiply totalLen by 8 to get the count in bits.
I had been adding the length of the message in bytes to the end of the
padding, but I think the RFC wants the length in bits, oops.
Unfortunately, it looks like we are still getting the same hashes as
before which indicates we aren't even using the length in the digestion?